requirements.txt
numpy~=1.19.1
scipy~=1.5.2
Pillow~=7.2.0

sklearn~=0.0
scikit-learn~=0.23.2
torch~=1.6.0

pip~=20.2.1
wheel~=0.34.2
setuptools~=49.2.1
torchvision~=0.7.0
